p2p: Connecting to a peer.
Very basic implementation of the P2PDevice.Connect.
For now, it's not possible create (or connect to) a group.
It's just possible connect directly to another peer,
as the go_intent parameter is being hard coded to the value 7,
each peer has 50% to be the group owner.

Originally authored by Flavio Ceolin <flavio.ceolin@profusion.mobi>